pyproject.toml
[tool.poetry]
description = ""
authors = ["AWS Prototyping ASEAN <aws-pace-asean@amazon.com>"]
readme = "README.md"
packages = [{include = "lib/backend"}]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11.0"
taskipy = "^1.10.3"

[tool.poetry.group.dev.dependencies]
mypy = "^1.0.0"
black = "^23.1.0"
pytest = "7.1.2"
pep8-naming = "^0.13.3"
pytest-mock = "^3.10.0"
pytest-cov = "^4.0.0"
ruff = "^0.0.291"

[tool.poetry.group.lambda.dependencies]
langchain = "^0.2.3"
numpy = "^1.26.4"
pgvector = "^0.2.5"
pg8000 = "^1.30.5"
boto3 = "^1.34.143"
aws-lambda-powertools = "^2.38.1"
requests = "^2.32.0"
PyJWT = "^2.8.0"
cryptography = "^36.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 132

[tool.mypy]
show_error_context = true
show_column_numbers = true
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true

[tool.ruff]
ignore = []
line-length = 132
select = [
    "E",
    "F",
    "W",
    "I",
    "B"
]

[tool.taskipy.tasks]
fmt_lint = { cmd = "task fmt && task lint", help = "format and lint" }
test = { cmd = "pytest -s -vv --cov=. --cov-branch", help = "runs all unit tests" }
fmt = { cmd = "task fmt_black && task fmt_isort", help = "format code" }
lint = { cmd = "task lint_black && task lint_ruff", help = "exec lint" }
fmt_black = "black lib/backend"
fmt_isort = "ruff check lib/backend --select I --fix-only --exit-zero"
lint_ruff = "ruff check lib/backend"
lint_mypy = "mypy lib/backend"
lint_black = "black --check lib/backend"